Changed CommandFilter, added ScannerRequest, removed FilesForRecord and LineRecorder as per comments
