+++ Record.txt
@@ -1,2 +1,4 @@
 1 line
 2 line
+3 line
+last line
